+++ NEWS.md
@@ -16,8 +16,10 @@
   parameter as a runtime parameter.
 * Most methods that takes output iterators, (including pointers that used in
   this fashion), now returns an iterator in the same way as `std::copy` etc.
-* `Allocator` is now always an alias to `AlignedAllocator`. The alignment is 32
-  for scalar types and `max(alignof(T), 16)` for others.
+* `AlignedAllocator` is renamed to `Allocator`. The old type alias `Allocator`
+  is removed. The default alignment is `AlignmentTrait<T>::value`, which is
+ `VSMC_ALIGNMENT` for scalar types and `max(alignof(T), VSMC_ALIGNMENT_MIN)`
+ for others.
 * `Vector` is now defined to be `std::vector<T, Allocator<T>>`
 * HDF5 IO functions' `append` parameter no longer has a default argument
 * `hdf5store` is now overloaded for `Monitor`